+++ nemo_cmd/deflate.py
@@ -19,7 +19,11 @@ Deflate variables in netCDF files using Lempel-Ziv compression.
 import logging
 import math
 import multiprocessing
-from pathlib import Path
+try:
+    from pathlib import Path
+except ImportError:
+    # Python 2.7
+    from pathlib2 import Path
 import shlex
 import subprocess
 import time